// src/generated/prometheus/prometheus_remote_write.ts
export interface WriteRequest {
  /**
   * @generated from protobuf field: repeated prometheus.TimeSeries timeseries = 1;
   */
  timeseries: TimeSeries[];
}
/**
 * @generated from protobuf message prometheus.TimeSeries
 */
export interface TimeSeries {
  /**
   * @generated from protobuf field: repeated prometheus.Label labels = 1;
   */
  labels: Label[];
  /**
   * @generated from protobuf field: repeated prometheus.Sample samples = 2;
   */
  samples: Sample[];
}
/**
 * @generated from protobuf message prometheus.Label
 */
export interface Label {
  /**
   * @generated from protobuf field: string name = 1;
   */
  name: string;
  /**
   * @generated from protobuf field: string value = 2;
   */
  value: string;
}
/**
 * @generated from protobuf message prometheus.Sample
 */
export interface Sample {
  /**
   * @generated from protobuf field: double value = 1;
   */
  value: number;
  /**
   * @generated from protobuf field: int64 timestamp = 2;
   */
  timestamp: bigint;
}
// @generated message type with reflection information, may provide speed optimized methods
class WriteRequest$Type extends MessageType<WriteRequest> {
  constructor() {
    super('prometheus.WriteRequest', [
      {
        no: 1,
        name: 'timeseries',
        kind: 'message',
        repeat: 2 /*RepeatType.UNPACKED*/,
        T: () => TimeSeries,
      },
    ]);
  }
  create(value?: PartialMessage<WriteRequest>): WriteRequest {
    const message = {timeseries: []};
    globalThis.Object.defineProperty(message, MESSAGE_TYPE, {enumerable: false, value: this});
    if (value !== undefined) reflectionMergePartial<WriteRequest>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: WriteRequest,
  ): WriteRequest {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* repeated prometheus.TimeSeries timeseries */ 1:
          message.timeseries.push(TimeSeries.internalBinaryRead(reader, reader.uint32(), options));
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: WriteRequest,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* repeated prometheus.TimeSeries timeseries = 1; */
    for (let i = 0; i < message.timeseries.length; i++)
      TimeSeries.internalBinaryWrite(
        message.timeseries[i],
        writer.tag(1, WireType.LengthDelimited).fork(),
        options,
      ).join();
    let u = options.writeUnknownFields;
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message prometheus.WriteRequest
 */
export const WriteRequest = new WriteRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class TimeSeries$Type extends MessageType<TimeSeries> {
  constructor() {
    super('prometheus.TimeSeries', [
      {no: 1, name: 'labels', kind: 'message', repeat: 2 /*RepeatType.UNPACKED*/, T: () => Label},
      {no: 2, name: 'samples', kind: 'message', repeat: 2 /*RepeatType.UNPACKED*/, T: () => Sample},
    ]);
  }
  create(value?: PartialMessage<TimeSeries>): TimeSeries {
    const message = {labels: [], samples: []};
    globalThis.Object.defineProperty(message, MESSAGE_TYPE, {enumerable: false, value: this});
    if (value !== undefined) reflectionMergePartial<TimeSeries>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: TimeSeries,
  ): TimeSeries {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* repeated prometheus.Label labels */ 1:
          message.labels.push(Label.internalBinaryRead(reader, reader.uint32(), options));
          break;
        case /* repeated prometheus.Sample samples */ 2:
          message.samples.push(Sample.internalBinaryRead(reader, reader.uint32(), options));
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: TimeSeries,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* repeated prometheus.Label labels = 1; */
    for (let i = 0; i < message.labels.length; i++)
      Label.internalBinaryWrite(
        message.labels[i],
        writer.tag(1, WireType.LengthDelimited).fork(),
        options,
      ).join();
    /* repeated prometheus.Sample samples = 2; */
    for (let i = 0; i < message.samples.length; i++)
      Sample.internalBinaryWrite(
        message.samples[i],
        writer.tag(2, WireType.LengthDelimited).fork(),
        options,
      ).join();
    let u = options.writeUnknownFields;
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message prometheus.TimeSeries
 */
export const TimeSeries = new TimeSeries$Type();
// ...
export const Label = new Label$Type();
// ...
export const Sample = new Sample$Type();